Fixed issue with VP HAS A where there was an out of bounds exception. Issue stemmed from a config file using only one "fragment" (aka target table). This made it impossible for a foreign key to exist between two tables - since there was only one table. Put in a check to ensure that the minimum number of fragments is two to avoid this issue in the future.
Fixed issue in MERGE ADD where there was infinite looping because of issues with the random selection of arguments. Added a "max_tries" counter which ensures that each argument slot of the skolem is only attempted to be filled 10 times before moving on to the next slot. There is, of course, a check to ensure that we don't end up with no skolem arguments being added.



git-svn-id: https://dblab.cs.toronto.edu/svn/boriscode/STBenchmark/trunk@422 de4125fd-08af-4ef3-add0-b7c1c68fae66